+++ src/pages/Login.tsx
@@ -1,7 +1,8 @@
 import AuthPage from "./Auth";
+import { LOGIN_MODE } from "../constants/auth";
 
 const Login = () => {
-  return <AuthPage mode="Login"/>;
+  return <AuthPage mode={LOGIN_MODE} />;
 };
 
 export default Login;
