+++ src/pages/Signup.tsx
@@ -1,5 +1,5 @@
 import AuthPage from "./Auth";
-
+import { SIGNUP_MODE } from "../constants/auth";
 const Signup = () => {
   return <AuthPage mode="Sign Up" />;
 };
