+++ src/utils/interface/user-data.ts
@@ -1,0 +1,6 @@
+export interface UserData {
+  id: string;
+  name: string;
+  email: string;
+  password: string;
+}
