+++ src/utils/types/board-view.ts
@@ -1,0 +1,4 @@
+export type ColumnItem = {
+  id: string;
+  title: string;
+};
