+++ src/utils/validation.ts
@@ -26,6 +26,7 @@ export function validatePassword(password: string): string | null {
   }
 }
 
+
 export function validateEmail(email: string): string | null {
   const emailPattern = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
